Make piggy bank total print to screen successfully

+++ src/piggyBank/AbstractMoney.java
@@ -2,6 +2,6 @@ package piggyBank;
 
 public abstract class AbstractMoney
 {
-	public abstract int getTotalValue();
+	public abstract double getTotalValue();
 	public abstract String getTotalCoins();
 }
